Support v5 PLIC-SW device for IPI
This commit implement the Andes V5 IPI mechanism. The different thing
with Sifive is that Andes use PLIC to handle IPI, and Sifive use CLINT
to handle IPI.

PLIC IPI pending array mapping as following:

Pending array start address: base + 0x1000
-------------------------------------
| core 3 | core 2 | core 1 | core 0 |
-------------------------------------
Each core X can send IPI to another core by setting the correspending
bit
in core X own region(see the below).

In each core region:
-----------------------------------------------
| bit 7 | bit 6 | bit 5 | bit 4 | ... | bit 0 |
-----------------------------------------------
The bit 7 is used to send IPI to core 0
The bit 6 is used to send IPI to core 1
The bit 5 is used to send IPI to core 2
The bit 4 is used to send IPI to core 3