Enable cache through dts information
Only main hart would parse dtb, the other harts assume always have L1
cache and try to enable L1 cache during initialization phase.

We don't check the next-level-cache property in dts here, because we
view L2 cache as a device, so if there is L2 node in dts, we would
enable it.